Revert "Add latitude and longitude data"
