+++ .gitignore
@@ -33,5 +33,3 @@ metals.sbt
 
 # data scrapped
 data.csv
-raw*.csv
-enhanced*.csv
