+++ src/main/scala/Main.scala
@@ -1,14 +1,9 @@
-import ReportRepresentation.{Report, ReportEnhanced, Table, Data}
 import com.github.nscala_time.time.Imports.*
 import net.ruippeixotog.scalascraper.browser.JsoupBrowser
 import net.sourceforge.htmlunit.cyberneko.HTMLElements.ElementList
 
-import scala.concurrent.duration.Duration
 import java.nio.charset.StandardCharsets
 import java.nio.file.{Files, Paths}
-import scala.concurrent.{Await, Future}
-import concurrent.ExecutionContext.Implicits.global
-import scala.util.{Failure, Success}
 
 @main def main(): Unit = downloadAndSaveReports()
 
@@ -16,29 +11,15 @@ private def downloadAndSaveReports(): Unit = {
   import scala.language.implicitConversions
   given JsoupBrowser()
 
+  // Scrapping data
   beautifyPrintln("Scrapping...")
   val reports = Scrapper.getReports
-  println(f"scrapped ${reports.length} reports")
+  println("done.")
 
-  beautifyPrintln("Saving and cleaning...")
-  val savingRawF = saveTable(reports, "./raw_data.csv")
-  savingRawF.onComplete {
-    case Failure(exception) => println(f"Could not save raw data: $exception")
-    case Success(_)         => println("Raw data saved inside ./raw.csv")
-  }
-  val savingEnhancedF =
-    Standardiser(reports)
-      .map(_.map(rep => ReportEnhanced.fromReport(rep)))
-      .map(standardized => saveTable(standardized, "./enhanced.csv"))
-  savingEnhancedF.onComplete {
-    case Failure(exception) => println(f"Could not save enhanced data: $exception")
-    case Success(_)         => println("Enhanced data saved inside ./enhanced_data.csv")
-  }
-  Await.result(savingRawF zip savingEnhancedF, Duration.Inf)
-}
-
-private def saveTable[A <: Data](reports: Table[A], filename: String): Future[Unit] = Future {
-  Files.write(Paths.get(filename), reports.toCSVFormat.getBytes(StandardCharsets.UTF_8))
+  // Saving data
+  beautifyPrintln("Saving data...")
+  Files.write(Paths.get("./data.csv"), reports.toCSVFormat.getBytes(StandardCharsets.UTF_8))
+  println("done.")
 }
 
 private def beautifyPrintln(str: String): Unit = println(
