+++ src/main/scala/Scrapper.scala
@@ -1,4 +1,3 @@
-import ReportRepresentation.{Report, Table, TableError}
 import net.ruippeixotog.scalascraper.browser.JsoupBrowser
 import net.ruippeixotog.scalascraper.browser.JsoupBrowser.JsoupElement
 import net.ruippeixotog.scalascraper.dsl.DSL.*
@@ -8,7 +7,7 @@ import net.ruippeixotog.scalascraper.model.*
 
 object Scrapper {
 
-  def getReports(using browser: JsoupBrowser): Table[Report] = {
+  def getReports(using browser: JsoupBrowser): Reports = {
     val nonConcatenatedRecords = allPagesSortedPerDate
       .map(downloadReportsFromPage)
       .zipWithIndex
@@ -20,8 +19,8 @@ object Scrapper {
       }
       .filter(_.nonEmpty)
       .flatten
-    val concatenatedRecords: Table[Report] =
-      nonConcatenatedRecords.foldLeft(Table.getEmpty) { (acc, report) =>
+    val concatenatedRecords: Reports =
+      nonConcatenatedRecords.foldLeft(Reports.getEmpty) { (acc, report) =>
         acc.copy(report.columns, acc.fields ++ report.fields)
       }
     concatenatedRecords
@@ -36,12 +35,12 @@ object Scrapper {
 
   private def downloadReportsFromPage(
       url: String
-  )(using browser: JsoupBrowser): Either[TableError, Table[Report]] = {
+  )(using browser: JsoupBrowser): Either[ReportsError, Reports] = {
     val dataFromUrlDoc = browser.get(url)
     val columns: List[JsoupElement] =
       (dataFromUrlDoc >> elementList("thead tr th font")).map(_.asInstanceOf[JsoupElement])
     val data: List[JsoupElement] =
       (dataFromUrlDoc >> elementList("tbody tr")).map(_.asInstanceOf[JsoupElement])
-    Table(columns.map(_.innerHtml), data.map(Report.fromJSoup))
+    Reports(columns.map(_.innerHtml), data.map(Report.fromJSoup))
   }
 }
